Actions documentation; Changed build to remove directory creation not working on Actions

+++ build-rpm.sh
@@ -51,15 +51,20 @@ if [ ! -z "$HTTP_PROXY" ] || [ ! -z "$http_proxy" ]; then
         sed s/\\[main\\]/\[main\]\\\nproxy=$TEMP_PROXY/g /tmp/$MOCK_CONFIG.cfg > /etc/mock/$MOCK_CONFIG.cfg
 fi
 
-OUTPUT_FOLDER=${OUTPUT_FOLDER}/${MOCK_CONFIG}
-if [ ! -d "$OUTPUT_FOLDER" ]; then
-        mkdir -p $OUTPUT_FOLDER
-else
-        rm -f $OUTPUT_FOLDER/*
-fi
+if [ ! -z "$GITHUB_BUILD" ]; then
+        OUTPUT_FOLDER=${OUTPUT_FOLDER}/${MOCK_CONFIG}
+        if [ ! -d "$OUTPUT_FOLDER" ]; then
+                mkdir -p $OUTPUT_FOLDER
+        else
+                rm -f $OUTPUT_FOLDER/*
+        fi
 
-if [ ! -d "$CACHE_FOLDER" ]; then
-        mkdir -p $CACHE_FOLDER
+        if [ ! -d "$CACHE_FOLDER" ]; then
+                mkdir -p $CACHE_FOLDER
+        fi
+else
+        $OUTPUT_FOLDER = $MOUNT_POINT
+        $CACHE_FOLDER = $MOUNT_POINT
 fi
 
 echo "=> Building parameters:"
